[TST] Add tests and move options-related tests to a new module.
